Advanced Node API: Promises, Code cleanup

+++ Advanced-Node-API/src/services/v1/user.services.ts
@@ -61,14 +61,18 @@ export const getUsers = async (): Promise<any> =>
 export const deleteUser = async (req: Request): Promise<any> => {
   try {
     return await UserSchema.findByIdAndDelete(req.params.userId)
-      .then(async () => {
-        return await ProductSchema.deleteMany({ user: req.params.userId })
-          .then((data: any) => {
-            return data;
-          })
-          .catch((error: Error) => {
-            throw error;
-          });
+      .then(async (data: any) => {
+        if (data === null) {
+          return data;
+        } else {
+          return await ProductSchema.deleteMany({ user: req.params.userId })
+            .then((data: any) => {
+              return data;
+            })
+            .catch((error: Error) => {
+              throw error;
+            });
+        }
       })
       .catch((error: Error) => {
         throw error;
